refactor(core-utils): simplify recursive schema helpers
Extract recursive schema property generation, recursive type alias rendering,
and local schema reference parsing into focused helpers so file-generators.ts
returns to orchestration. Remove unused recursive context state and add direct
coverage for shared schema reference parsing.
